requirements.txt
streamlit==1.35.0
streamlit-webrtc
opencv-python==4.5.5.64
opencv-contrib-python==4.5.5.64
matplotlib
plotly
numpy
pillow
